using TurtleDB for websocket client
